Move topic store health collection
