+++ packages/column-live-view-engine/src/topic-store.ts
@@ -1,12 +1,4 @@
-import { Effect } from "effect";
-import { activeStoreRawQueryExecutionCount } from "./active-query";
-import { collectTopicStoreHealthView, type TopicStoreHealthState } from "./topic-store-health";
-import {
-  TopicStore,
-  topicStoreRawQueryMetadata,
-  topicStoreReadModel,
-  topicStoreState,
-} from "./topic-store-state";
+import { TopicStore, topicStoreRawQueryMetadata, topicStoreReadModel } from "./topic-store-state";
 
 export { TopicStore, topicStoreRawQueryMetadata, topicStoreReadModel };
 export {
@@ -33,21 +25,5 @@ export {
   releaseTopicStoreMaterializedQueryExecution,
   releaseTopicStoreRawQueryExecution,
 } from "./topic-store-query";
+export { collectTopicStoreHealth } from "./topic-store-health";
 export type { TopicStoreSubscriptionPermit } from "./topic-store-state";
-
-const topicStoreHealthState = (store: TopicStore, activeViews: number): TopicStoreHealthState => {
-  const state = topicStoreState(store);
-  return {
-    activeViews,
-    healthLedger: state.healthLedger,
-    subscribers: state.subscribers,
-    topic: store.topic,
-  };
-};
-
-export const collectTopicStoreHealth = Effect.fn("ColumnLiveViewEngine.topicStore.health")(
-  function* (store: TopicStore, closed: boolean) {
-    const activeViews = yield* activeStoreRawQueryExecutionCount(topicStoreReadModel(store));
-    return yield* collectTopicStoreHealthView(topicStoreHealthState(store, activeViews), closed);
-  },
-);
